The HoldResponse model isn't Sierra specific
