+++ common/stacks/src/test/scala/uk/ac/wellcome/platform/api/common/services/SierraServiceTest.scala
@@ -8,7 +8,7 @@ import org.scalatest.EitherValues
 import uk.ac.wellcome.platform.api.common.fixtures.ServicesFixture
 import uk.ac.wellcome.platform.api.common.models._
 import com.github.tomakehurst.wiremock.client.WireMock._
-import weco.api.stacks.models.StacksUserIdentifier
+import weco.api.stacks.models.{HoldRejected, StacksUserIdentifier}
 import weco.catalogue.internal_model.identifiers.IdentifierType.SierraSystemNumber
 import weco.catalogue.internal_model.identifiers.SourceIdentifier
